requirements.txt
chromadb
boto3
openai
tiktoken
beautifulsoup4
requests
fake-useragent
python-dateutil